add check in build script for valid package.json

+++ build.sh
@@ -27,5 +27,12 @@ echo "Version: " $VERSION "->" $NEW_VERSION
 ################################################
 
 jq '.version = "'$NEW_VERSION'"' pkg/package.json > pkg/package.temp.json
+if [ ! -s pkg/package.temp.json ]; then
+   echo "ERROR: 'pkg/package.temp.json' was generated empty."
+   echo "- Did you install jq? (brew install jq)"
+   echo "- Does 'pkg/package.json' exist? Interrupting a build may cause it to be lost. Discard git changes to restore it."
+   echo "- Is your 'pkg/package.json' valid json?"
+   exit 1
+fi
 wasm-pack build --target web
 mv pkg/package.temp.json pkg/package.json
